requirement.txt
numpy
tensorflow
streamlit
pillow
os